created security config, custom classes for OAuth2 Authenticationand custom OAuth2 user service
the security config contains the security settings for oauth2 socail login, as well as email and password based login. The custom class for authentication oauth2, serves to compare the value of the state in the authorization request. The class provides functionality to store the authorization request in cookies and retrieve it. Finally, we have the custom oauth2 user service, used when the access token is obtained from the oauth2 provider, where the data will be registered, updated, etc.